:recycle: refactor: Rename Riddle to Gate
